Add standard constructors to exceptions
Fixes #1207
