+++ .editorconfig
@@ -68,5 +68,8 @@ dotnet_style_prefer_is_null_check_over_reference_equality_method = false:silent
 dotnet_style_predefined_type_for_locals_parameters_members = true:silent
 dotnet_style_predefined_type_for_member_access = true:silent
 
+# CA1032: Implement standard exception constructors
+dotnet_diagnostic.CA1032.severity = warning
+
 # CA2000: Dispose objects before losing scope
-dotnet_diagnostic.CA2000.severity = suggestion
+dotnet_diagnostic.CA2000.severity = warning
