+++ SteamKit2/SteamKit2/Steam/Authentication/AuthenticationException.cs
@@ -5,7 +5,6 @@ namespace SteamKit2.Authentication
     /// <summary>
     /// Thrown when <see cref="SteamAuthentication"/> fails to authenticate.
     /// </summary>
-    [Serializable]
     public sealed class AuthenticationException : Exception
     {
         /// <summary>
@@ -13,11 +12,19 @@ namespace SteamKit2.Authentication
         /// </summary>
         public EResult Result { get; }
 
-        /// <summary>
-        /// Initializes a new instance of the <see cref="AuthenticationException"/> class.
-        /// </summary>
+        /// <inheritdoc/>
         public AuthenticationException()
-        { 
+        {
+        }
+
+        /// <inheritdoc/>
+        public AuthenticationException(string message) : base(message)
+        {
+        }
+
+        /// <inheritdoc/>
+        public AuthenticationException(string message, Exception innerException) : base(message, innerException)
+        {
         }
 
         /// <summary>
@@ -25,8 +32,8 @@ namespace SteamKit2.Authentication
         /// </summary>
         /// <param name="message">The message that describes the error.</param>
         /// <param name="result">The result code that describes the error.</param>
-        public AuthenticationException( string message, EResult result )
-            : base( $"{message} with result {result}." )
+        public AuthenticationException(string message, EResult result)
+            : base($"{message} with result {result}.")
         {
             Result = result;
         }
